pyproject.toml
[project]
name = "8-fold-assistant"
version = "0.1.0"
description = "AI-powered 8-fold Assistant"
authors = [
    { name = "Jnanesh P S" }
]

requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "httpx>=0.28.1",
    "mem0ai>=1.0.1",
    "openai>=2.8.1",
    "openai-agents>=0.6.1",
    "perplexityai>=0.20.0",
    "psycopg[binary]>=3.2.13",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
]

[project.scripts]
start = "app:main"

[tool.uv]
package = true

[tool.setuptools]
packages = ["backend", "ui"]
